requirements.txt
django~=5.0.4
django-debug-toolbar~=4.3.0